Fixing a high-importance security issue
An issue in validating administrator accounts by domains could be abused to leak private info with an easy attack vector.
This commit should fix the issue, but we should work on improving admin account security.

We have strong beliefs that this issue wasn't abused and that we have caught the issue in time.